chore(1-3316): cleanup
